perf/core: Add PERF_AUX_FLAG_COLLISION to report colliding samples
The ARM SPE architecture permits an implementation to ignore a sample
if the sample is due to be taken whilst another sample is already being
produced. In this case, it is desirable to report the collision to
userspace, as they may want to lower the sample period.

This patch adds a PERF_AUX_FLAG_COLLISION flag, so that such events can
be relayed to userspace.

Signed-off-by: Will Deacon <will.deacon@arm.com>
merged with upstream PERF_AUX_FLAG_PARTIAL, changed value to next power of 2 (unless was supposed to match both TRUNCATED and OVERWRITE?)
Signed-off-by: Kim Phillips <kim.phillips@arm.com>