perf tools: Add data block processing function
The trace decoder library, after it has been initialized, works by being
repeatedly called to process data and given a pointer into a data buffer.
Each call consumes 0 or more of the buffer, and may result in one or more
callbacks to the trace packet handler registered with the decoder. Multiple
calls may occur as a trace stream atom can encode multiple branch events,
and thus generate multiple trace packets in the decoded trace. The function
returns a result code that determines whether it needs to be flushed or
decoding can continue (up to the point at which the data buffer has been
exhausted).

Signed-off-by: Tor Jeremiassen <tor@ti.com>