perf tools: Add channel context item to track packet sources
This patch adds a list structure to the decoder data to allow each decoded
CoreSight ID to be associated with the callback context.

This allows the tracking of the source of the packets in the
perf report -D command.

The code was contributed by Mike Leach in a separate commit which had to
be adjusted and distributed across several patches.

Cc: Mike Leach <mike.leach@linaro.org>
Signed-off-by: Tor Jeremiassen <tor@ti.com>