perf tools: Add trace packet print for dump_trace option
perf report may be called with the -D option to dump the raw trace in
ASCII.

This patch adds two functions. One to print the string representation of a
trace packet. The other takes an auxtrace buffer and instantiates a trace
decoder configured only to parse the trace stream into trace packets and
print the packets using the first function. Then pushes the trace data
through the instantiated decoder to generate the output.

Signed-off-by: Tor Jeremiassen <tor@ti.com>