perf tools: Add function to read data from dsos
The trace decoder library requires access to the instruction opcodes of the
program that was traced in order to be able to decode the trace stream into
a series of executed instruction ranges.

This patch implements a function that will be used by the trace decoder
interface to enable the trace decoder library to access instruction opcodes
stored in DSOs.

Signed-off-by: Tor Jeremiassen <tor@ti.com>