perf tools: Add functions to allocate and free queues
Auxtrace events are associated with auxtrace_queues. The auxtrace queues
are extended using the "priv" field to point to a cs-etm specific
structure associated with each queue.

This patch adds functions to allocate and free these structures, including
instantiating the trace decoder library interface, and providing the
function used to register a memory access function with the decoder
library.

Signed-off-by: Tor Jeremiassen <tor@ti.com>