perf tools: Add functions for full etmv4i packet decode
The full decode of a coresight trace requires the decoder library be
properly configured, with a "generic" packet handler function. The packets
are emitted into a buffer with a delay so as to be able to add additional
information to the most recently decoded packet to encode information
about taking/returning from interrupts/exceptions.

This patch adds the code to configure the trace decoder for full packet
decode, provides the function to handle the trace packets, buffer, and
annotate them when necessary.

Signed-off-by: Tor Jeremiassen <tor@ti.com>