perf annotate: Don't prepend symfs path to vmlinux path
When selecting the filename to pass to objdump for disassembly, if there is
no entry for vmlinux in the build-id cache then the fallback code prepends
symfs path to vmlinux path. If the vmlinux option was specified, this is
unnecessary and, as a result, vmlinux is not found.

Don't prepend symfs path if DSO is [kernel.kallsyms]

Signed-off-by: Martin Vuille <jpmv27@aim.com>